Fix #2040483 ("AppArmor denies crun sending signals to containers (stop, kill)")
This makes projects using AppArmor bits from
golang-github-containers-common (notably podman) work with AppArmor
v4.0.0.

There is a similar issue with containerd clients and docker. The fix was
merged to the containerd upstream[1]. The fix to moby (docker) was
submitted but seems to have stalled[2]. Upstream notes we should fix
regressions we introduced in Ubuntu or perhaps at least introduce a
generic way to refer to OCI runtimes under a single peer name. I suspect
we would get similar objections in containers/common. That's why I
haven't yet submitted the patch to the upstream.

In the meantime, patch this library so that podman can work with OCI
runtimes we currently confine.

[1] https://github.com/containerd/containerd/pull/10123
[2] https://github.com/moby/moby/pull/47749

Bug: https://bugs.launchpad.net/ubuntu/+source/libpod/+bug/2040483